additions for sparlq11 based on http://www.w3.org/2009/sparql/docs/tests/as of 2012-03-23

+++ spec/w3c/data-r2/type-promotion/type-promotion-22-native_spec.rb
@@ -60,7 +60,7 @@ PREFIX xsd: <http://www.w3.org/2001/XMLSchema#>
 ASK
  WHERE { t:short1 rdf:value ?l .
          t:decimal1 rdf:value ?r .
-         FILTER ( datatype(?l + ?r) = xsd:integer ) }
+         FILTER ( datatype(?l + ?r) = xsd:decimal ) }
 
 }
     end
